fix another logic bug: resting on other sneks is ok
Previously if we were about to fall into a spike, only fruit or solid
ground could save us. But this is not enough - resting on another snek
that is grounded is fine too. Now we apply the spike check at the end,
after taking into account other sneks that we might be resting on.