add "doom" tile
We don't really have these tiles in the game, but they are useful to cut
down the search space for levels with crazy combinatorial explosion. We
often can manually reason which parts of the level no sneks or bloks
will touch. This, for example, allows us to solve level 40, which I
wasn't able to solve before.